fix(striker-ui): add external link to MITM wiki in ManageChangedSSHKeysForm
