+++ striker-ui/types/ManageChangedSSHKeysForm.d.ts
@@ -1,0 +1,14 @@
+type ChangedSSHKeys = {
+  [hostUUID: string]: {
+    hostName: string;
+    hostUUID: string;
+    ipAddress: string;
+    isChecked?: boolean;
+  };
+};
+
+type ManageChangedSSHKeysFormOptionalProps = {
+  mitmExternalHref?: LinkProps['href'];
+};
+
+type ManageChangedSSHKeysFormProps = ManageChangedSSHKeysFormOptionalProps;
